backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
sqlalchemy==2.0.*
python-dotenv==1.1.*
groq>=0.12.0
pydantic==2.10.*
